+++ crypto-contract-value/src/exchanges/bybit.rs
@@ -1,6 +1,6 @@
 pub use crypto_market_type::MarketType;
 
-pub(crate) fn get_contract_value(market_type: MarketType, _pair: &str) -> Option<f32> {
+pub(crate) fn get_contract_value(market_type: MarketType, _pair: &str) -> Option<f64> {
     match market_type {
         // Each inverse contract value is 1 USD, see:
         // https://www.bybit.com/data/basic/inverse/contract-detail?symbol=BTCUSD
